+++ fourtooncookie/src/main/java/com/startingblue/fourtooncookie/member/MemberController.java
@@ -1,4 +1,4 @@
-package com.startingblue.fourtooncookie.member.domain;
+package com.startingblue.fourtooncookie.member;
 
 import com.startingblue.fourtooncookie.member.dto.response.MemberSavedResponse;
 import com.startingblue.fourtooncookie.member.dto.request.MemberUpdateRequest;
