--- fourtooncookie/src/main/java/com/startingblue/fourtooncookie/member/authorization/exception/InvalidPathVariableException.java
@@ -1,8 +1,0 @@
-package com.startingblue.fourtooncookie.member.authorization.exception;
-
-public class InvalidPathVariableException extends IllegalArgumentException {
-
-    public InvalidPathVariableException(String message) {
-        super(message);
-    }
-}
