--- fourtooncookie/src/main/java/com/startingblue/fourtooncookie/member/authorization/exception/TokenNotFoundException.java
@@ -1,10 +1,0 @@
-package com.startingblue.fourtooncookie.member.authorization.exception;
-
-import java.util.NoSuchElementException;
-
-public class TokenNotFoundException extends NoSuchElementException {
-
-    public TokenNotFoundException(String message) {
-        super(message);
-    }
-}
