--- fourtooncookie/src/main/java/com/startingblue/fourtooncookie/member/authorization/exception/UnauthorizedAccessException.java
@@ -1,8 +1,0 @@
-package com.startingblue.fourtooncookie.member.authorization.exception;
-
-public class UnauthorizedAccessException extends RuntimeException {
-
-    public UnauthorizedAccessException(String message) {
-        super(message);
-    }
-}
